Implement download_local
The `p4 print` command allows us to retrieve file contents without
requiring a workspace. Since we can specify a Perforce label as part of
the command (not implemented yet), we can download directly to the
install directory instead of going through an intermediate zip/extract
step to get the correct version.